MEDIUM: clock: change the type of 'now' to a ullong with 32:32
Now the 'now' variable is of the same type as global_now: a single
ullong made of 2*32 bit values (sec and usec) that can be accessed
atomically if needed. The real purpose is to make sure nobody uses
now by accident to report a date to a human anymore. Some improvements
are clearly deserved so that timeval progressively goes away from the
intermediary operations, and maybe now/global_now switch to a u64
representing nanoseconds (585 years before rollover, that's OK). It
looks like this shouldn't be difficult anymore as we're already
close to having a linear time scale.