MINOR: clock: temporarily replace now with CLOCK_TO_TV(now)
This is needed because there are quite a number of places where we use
&now in tv_ms_elapsed() to calculate an elapsed time since a previous
event that was copied from now into a timeval variable. Since the API
uses timeval, we have to remain compatible with it for the time being,
until the rest is migrated away from timeval as well.

The reason that we do not use ticks for these is that historically these
timestamps were relative to the accept_date which was relative to the
wall clock time for logging purposes. However since integers are used
everywhere and the measurement units are milliseconds, none of these
timestamps supports durations longer than 32-bit anyway so it's not a
technical limitation at all, just a historical one. It would be finer
to just replace all timeval except accept_date with ticks and use
regular integer add/sub operations. This would also help migrating
towards microsecond later.